+++ lib/core/ui/cached_network_image_widget.dart
@@ -119,7 +119,7 @@ class _CachedNetworkImageState extends State<CachedNetworkImage> {
       final request = await client.getUrl(Uri.parse(widget.url));
       request.headers.set(
         HttpHeaders.userAgentHeader,
-        'ClipFlow/1.0.4 ImageCache',
+        'ClipFlow/1.0.5 ImageCache',
       );
       final response = await request.close();